Extracted popover into reusable component and reused it
