+++ src/features/amUI/common/AccessPackageInfoPopover/AccessPackageInfoPopover.module.css
@@ -1,0 +1,16 @@
+.infoBox {
+  padding: var(--ds-spacing-2);
+  display: flex;
+  flex-direction: column;
+  gap: var(--ds-spacing-2);
+}
+
+.popover {
+  min-width: 400px;
+}
+
+@media (max-width: 768px) {
+  .popover {
+    min-width: 100%;
+  }
+}
