+++ src/features/amUI/poaOverview/AccessPackagePermissions.module.css
@@ -2,20 +2,3 @@
   display: flex;
   align-items: center;
 }
-
-.infoBox {
-  padding: var(--ds-spacing-2);
-  display: flex;
-  flex-direction: column;
-  gap: var(--ds-spacing-2);
-}
-
-.popover {
-  min-width: 400px;
-}
-
-@media (max-width: 768px) {
-  .popover {
-    min-width: 100%;
-  }
-}
