+++ src/localizations/en.json
@@ -482,7 +482,8 @@
     "request_service": "Which service do you need?",
     "give_service_to_name": "Give <strong>{{name}}</strong> power of attorney for new service",
     "request_package": "Which access package do you need?",
-    "give_package_to_name": "Give <strong>{{name}}</strong> power of attorney for new access package",
+    "give_package_to_name": "What will <strong>{{name}}</strong> be working on?",
+    "package_delegation_description": "Open an area to see which access packages it contains.",
     "name_will_receive": "<strong>{{name}}</strong> will receive:",
     "name_has_the_following": "<strong>{{name}}</strong> has the following:",
     "name_requests_access_to": "<strong>{{name}}</strong> is asking for:",
