+++ src/localizations/no_nb.json
@@ -479,7 +479,8 @@
     "request_service": "Hvilken tjeneste vil du be om?",
     "give_service_to_name": "Gi <strong>{{name}}</strong> fullmakt til ny tjeneste",
     "request_package": "Hvilken tilgangspakke vil du be om?",
-    "give_package_to_name": "Gi <strong>{{name}}</strong> fullmakt til ny tilgangspakke",
+    "give_package_to_name": "Hva skal <strong>{{name}}</strong> jobbe med?",
+    "package_delegation_description": "Åpne et område for å se hvilke tilgangspakker det inneholder.",
     "name_will_receive": "<strong>{{name}}</strong> vil få:",
     "name_has_the_following": "<strong>{{name}}</strong> har:",
     "name_requests_access_to": "<strong>{{name}}</strong> ber om:",
