+++ src/localizations/no_nn.json
@@ -479,7 +479,8 @@
     "request_service": "Kva for teneste vil du be om?",
     "give_service_to_name": "Gi <strong>{{name}}</strong> fullmakt til ny teneste",
     "request_package": "Kva for tilgangspakke vil du be om?",
-    "give_package_to_name": "Gi <strong>{{name}}</strong> fullmakt til ny tilgangspakke",
+    "give_package_to_name": "Kva skal <strong>{{name}}</strong> jobbe med?",
+    "package_delegation_description": "Åpne eit område for å sjå kva tilgangspakker det inneheld.",
     "name_will_receive": "<strong>{{name}}</strong> vil få:",
     "name_has_the_following": "<strong>{{name}}</strong> har:",
     "name_requests_access_to": "<strong>{{name}}</strong> ber om:",
